chore: ドメインモデルのファイル整理
