+++ src/components/map/mapEventHook.ts
@@ -1,15 +1,15 @@
 import { MutableRefObject, useRef, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
 import { useSearchParams } from "react-router-dom"
+import { parseQueryBoolean } from "../../model/diagram"
+import { Line } from "../../model/line"
+import { LatLng } from "../../model/location"
+import { Station } from "../../model/station"
 import * as action from "../../redux/actions"
 import { selectMapState } from "../../redux/selector"
 import { AppDispatch } from "../../redux/store"
 import StationService from "../../script/StationService"
-import { Line } from "../../script/line"
-import { LatLng } from "../../script/location"
 import { logger } from "../../script/logger"
-import { Station } from "../../script/station"
-import { parseQueryBoolean } from "../../script/utils"
 import { useRefCallback } from "../hooks"
 import { NavType, isStationDialog } from "../navState"
