+++ src/components/map/mapHook.ts
@@ -1,14 +1,14 @@
 import { MutableRefObject, RefObject, useEffect, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
+import { PolylineProps, RectBounds, getBounds, getZoomProperty, isInsideRect } from "../../model/diagram"
+import { Line } from "../../model/line"
+import { LatLng, MapCenter } from "../../model/location"
+import { Station } from "../../model/station"
 import * as action from "../../redux/actions"
 import { selectMapState } from "../../redux/selector"
 import { AppDispatch } from "../../redux/store"
 import StationService, { DataType } from "../../script/StationService"
-import { Line } from "../../script/line"
-import { LatLng, MapCenter } from "../../script/location"
 import { logger } from "../../script/logger"
-import { Station } from "../../script/station"
-import { PolylineProps, RectBounds, getBounds, getZoomProperty, isInsideRect } from "../../script/utils"
 import { useRefCallback } from "../hooks"
 import { NavType, isStationDialog } from "../navState"
 import { useHighVoronoi } from "./voronoiHook"
