+++ src/model/station.ts
@@ -1,5 +1,5 @@
+import { logger } from "../script/logger"
 import { LatLng } from "./location"
-import { logger } from "./logger"
 
 export interface DelaunayStation {
   code: number
